+++ src/components/posts/PostList.js
@@ -6,41 +6,59 @@ import {PostByAuthor} from "./PostByAuthor"
 import "./Posts.css";
 import { getCategories } from "../../managers/categories";
 
-export const PostList = ({ token }) => {
+export const PostList = ({ token, authorChoice , selectedCategory }) => {
   const [posts, setPosts] = useState([]);
   const [filteredPosts, setFilteredPosts] = useState([]);
-  const [categories, setCategories] = useState([])
-  const [selectedCategory, setSelectedCategory] = useState(0)
   const navigate = useNavigate();
 
   useEffect(
     () => {
-      getCategories().then((categoryData) => setCategories(categoryData))
       getPosts().then((postData) => setPosts(postData))
+      setFilteredPosts(posts)
     }, [])
 
   useEffect(
-    () => {
-      if (selectedCategory === 0) { setFilteredPosts(posts) }
-      else {
-        const filteredCopy = posts.filter(post => post.category_id === parseInt(selectedCategory))
-        setFilteredPosts(filteredCopy)
-      }
-    },
-    [posts, selectedCategory]
-  )
+      () => {
+        if (selectedCategory === 0 && authorChoice === 0) { setFilteredPosts(posts) }
+        else if (selectedCategory !== 0 && authorChoice === 0) {
+          const filteredCopy = posts.filter(post => post.category_id === parseInt(selectedCategory))
+          setFilteredPosts(filteredCopy)
+        }
+        else if (selectedCategory === 0 && authorChoice !== 0) {
+          const filteredPostList = posts.filter(post => post.user_id === parseInt(authorChoice))
+          setFilteredPosts(filteredPostList)
+        }
+        else if (selectedCategory !== 0 && authorChoice !== 0) {
+          const filteredPostList = posts.filter(post => post.user_id === parseInt(authorChoice) && post.category_id === parseInt(selectedCategory))
+          setFilteredPosts(filteredPostList)
+        }
+      },
+      [posts, selectedCategory, authorChoice]
+    )
+
+//   useEffect(
+//     () => {
+//       if (selectedCategory === 0) { setFilteredPosts(posts) }
+//       else {
+//         const filteredCopy = posts.filter(post => post.category_id === parseInt(selectedCategory))
+//         setFilteredPosts(filteredCopy)
+//       }
+//     },
+//     [posts, selectedCategory]
+//   )
+
+//   useEffect(() => {
+//     if (authorChoice === 0) {
+//         setFilteredPosts(posts)
+//     } else {
+//         const filteredPostList = posts.filter(post => post.user_id === parseInt(authorChoice))
+//         setFilteredPosts(filteredPostList)
+//     }
+// }
+//     , [posts, authorChoice])
+    
   return (
     <>
-      <PostByAuthor posts={posts} /><br></br>
-      <label htmlFor="categories">Search By Category</label><br></br>
-        <select onChange={(event) => { setSelectedCategory(parseInt(event.target.value)) }}>
-          <option value="0" name="category_id" className="form-control" >View All</option>
-          {categories.map(category => (
-            <option key={`category--${category.id}`} value={category.id}>
-              {category.label}
-            </option>
-          ))}
-        </select>
         <div className="addPostButton">
           Add Post <button onClick={() => navigate("newpost")}>+</button>
         </div>
